fix: wrap rmdirSync with existsSync check to ensure works with node 12 & 16

+++ index.js
@@ -53,8 +53,10 @@ module.exports = class MakerSnap extends MakerBase {
 		log(`Snap file moved to: ${finalSnapLocation}`);
 
 		const snapcraftDirectory = path.dirname(snapFileLocation);
-		fse.rmdirSync(snapcraftDirectory, {recursive: true});
-		log(`Tidy up; snapcraft files deleted from: ${snapcraftDirectory}`);
+		if (fse.existsSync(snapcraftDirectory)) {
+			fse.rmdirSync(snapcraftDirectory, {recursive: true});
+			log(`Tidy up; snapcraft files deleted from: ${snapcraftDirectory}`);
+		}
 
 		log(`Finishing snap maker, passing back to electron-forge with: ${finalSnapLocation}`);
 		return [finalSnapLocation];
